+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # Peplum ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Added a system command for reversing the current sort order.
+  ([#46](https://github.com/davep/peplum/pull/46))
+
 ## v0.4.2
 
 **Released: 2025-03-17**